Split table logic
