+++ frontend/src/shipping_queue/tables/ShippingQueueTable.jsx
@@ -6,6 +6,8 @@ import { styled } from "@mui/system";
 import { createColumnFilters } from "../../utils/TableFilters";
 import { getCheckboxColumn } from "../../components/CheckboxColumn";
 import ShipQueuePackSlipDrowdown from "./ShipQueuePackSlipDropdown";
+import { API } from "../../services/server";
+import CreateShipmentDialog from "../../create_shipment/CreateShipmentDialog";
 
 const useStyle = makeStyles((theme) => ({
   root: {
@@ -45,17 +47,23 @@ const ShippingQueueDataGrid = styled(DataGrid)`
 `;
 
 const ShippingQueueTable = ({
+  shippingQueue,
   tableData,
-  onRowClick,
-  isSelectAllOn,
-  onSelectAll,
-  selectedCustomerId,
-  selectionOrderIds,
   sortModel,
   setSortModel,
+  selectedOrderIds,
+  setSelectedOrderIds,
+  onCreateShipmentClose,
+  setShippingQueue,
+  setFilteredShippingQueue,
+  createShipmentOpen,
+  currentDialogState,
+  setCurrentDialogState,
 }) => {
   const classes = useStyle();
   const [queueData, setQueueData] = useState(tableData);
+  const [selectedCustomerId, setSelectedCustomerId] = useState(null);
+  const [isSelectAllOn, setIsSelectAll] = useState(false);
 
   const numRowsPerPage = 10;
 
@@ -71,11 +79,121 @@ const ShippingQueueTable = ({
 
   const storedTableData = useMemo(() => tableData, [tableData]);
 
+  const reloadData = useCallback(() => {
+    async function fetchData() {
+      const data = await Promise.all([API.getShippingQueue()]);
+      return { queue: data[0] };
+    }
+
+    fetchData().then((data) => {
+      // Gather the queue data for the table
+      let queueTableData = [];
+      data?.queue?.packingSlips.forEach((e) => {
+        queueTableData.push({
+          id: e._id,
+          orderNumber: e.orderNumber,
+          packingSlipId: e.packingSlipId,
+          customer: e.customer,
+          items: e.items,
+        });
+      });
+
+      // The set state order is important
+      setSelectedCustomerId(null);
+      setSelectedOrderIds([]);
+      setShippingQueue(queueTableData);
+      setFilteredShippingQueue(queueTableData);
+    });
+  }, [setFilteredShippingQueue, setSelectedOrderIds, setShippingQueue]);
+
+  useEffect(() => {
+    reloadData();
+  }, [reloadData]);
+
+  const handleSelection = useCallback(
+    (selection, tableData) => {
+      let newSelection = selectedOrderIds;
+      if (selectedOrderIds.includes(selection)) {
+        // remove it
+        newSelection = selectedOrderIds.filter((e) => e !== selection);
+        // if something is deselected then selectAll is false
+        setIsSelectAll(false);
+      } else {
+        // add it
+        newSelection.push(selection);
+
+        // if the new selection contains all possible selected order numbers
+        // then select all is on
+        const selectedCustId = tableData?.find((e) => e.id === selection)
+          ?.customer._id;
+        const idsWithSelectedCustId = tableData
+          ?.filter((e) => e.customer._id === selectedCustId)
+          .map((e) => e.id);
+
+        setIsSelectAll(
+          idsWithSelectedCustId.sort().toString() ===
+            newSelection.sort().toString()
+        );
+      }
+      return newSelection;
+    },
+    [selectedOrderIds]
+  );
+
+  const onRowClick = useCallback(
+    (selectionModel, tableData) => {
+      const newSelectedOrderIds = handleSelection(selectionModel, tableData);
+      setSelectedOrderIds([...newSelectedOrderIds]);
+
+      setSelectedCustomerId(
+        tableData?.find(
+          (e) => newSelectedOrderIds.length > 0 && e.id === selectionModel
+        )?.customer?._id ?? null
+      );
+    },
+    [handleSelection, setSelectedOrderIds]
+  );
+
+  const onSelectAll = useCallback(
+    (value, tableData) => {
+      setIsSelectAll(value);
+
+      if (value) {
+        if (selectedOrderIds.length > 0) {
+          // Something is selected, so we need to select the remaining
+          // that matach selectedOrderNumber
+          setSelectedOrderIds(
+            tableData
+              .filter((e) => e.customer?._id === selectedCustomerId)
+              .map((e) => e.id)
+          );
+        } else if (selectedOrderIds.length === 0) {
+          // Nothing selected yet, so select the first row and all that match
+          // the first row order number
+
+          setSelectedOrderIds(
+            tableData
+              .filter((e) => e.customer?._id === tableData[0]?.customer?._id)
+              .map((e) => e.id)
+          );
+          setSelectedCustomerId(
+            tableData?.find((e) => e.id === tableData[0].id)?.customer?._id ??
+              null
+          );
+        }
+      } else {
+        setSelectedOrderIds([]);
+        setSelectedCustomerId(null);
+      }
+    },
+    [selectedOrderIds, selectedCustomerId, setSelectedOrderIds]
+  );
+
   const columns = useMemo(
     () => [
       getCheckboxColumn(
         isDisabled,
-        selectionOrderIds,
+        selectedOrderIds,
         isSelectAllOn,
         storedTableData,
         onSelectAll,
@@ -104,10 +222,16 @@ const ShippingQueueTable = ({
       isSelectAllOn,
       onRowClick,
       onSelectAll,
-      selectionOrderIds,
+      selectedOrderIds,
       storedTableData,
     ]
   );
+
+  const packingSlipIds = useMemo(() => {
+    return shippingQueue
+      .filter((e) => selectedOrderIds.includes(e.id))
+      .map((e) => e.id);
+  }, [shippingQueue, selectedOrderIds]);
 
   const sortDataByModel = useCallback(
     (model) => {
@@ -120,7 +244,7 @@ const ShippingQueueTable = ({
         // setQueueData(
         return clickedColumnField?.handler(
           model[0]?.sort,
-          selectionOrderIds,
+          selectedOrderIds,
           tableData
         );
         // );
@@ -128,7 +252,7 @@ const ShippingQueueTable = ({
         return tableData;
       }
     },
-    [columns, selectionOrderIds, tableData]
+    [columns, selectedOrderIds, tableData]
   );
 
   useEffect(() => {
@@ -136,10 +260,10 @@ const ShippingQueueTable = ({
     setQueueData(d);
   }, [sortModel, sortDataByModel]);
 
-  const [page, setPage] = useState(0)
+  const [page, setPage] = useState(0);
 
   const handlePageChange = (event, newPage) => {
-    setPage(newPage)
+    setPage(newPage);
   };
 
   const generateTablePagination = useCallback(() => {
@@ -188,15 +312,15 @@ const ShippingQueueTable = ({
         }}
         components={{
           Footer: () =>
-            selectionOrderIds.length > 0 ? (
+            selectedOrderIds.length > 0 ? (
               <Grid container alignItems="center" spacing={2}>
                 <Grid container item xs={6} justifyContent="flex-start">
                   <Typography sx={{ padding: "8px" }}>
-                    {selectionOrderIds.length} rows selected
+                    {selectedOrderIds.length} rows selected
                   </Typography>
                 </Grid>
                 <Grid container item xs={6} justifyContent="flex-end">
-                {generateTablePagination()}
+                  {generateTablePagination()}
                 </Grid>
               </Grid>
             ) : (
@@ -205,6 +329,29 @@ const ShippingQueueTable = ({
               </Grid>
             ),
         }}
+      />
+
+      <CreateShipmentDialog
+        customer={
+          shippingQueue.filter((e) => selectedOrderIds.includes(e.id))[0]
+            ?.customer
+        }
+        packingSlipIds={packingSlipIds}
+        open={createShipmentOpen}
+        onClose={onCreateShipmentClose}
+        currentState={currentDialogState}
+        setCurrentState={setCurrentDialogState}
+        parts={shippingQueue
+          .filter((e) => selectedOrderIds.includes(e.id))
+          .reduce(
+            (result, current) =>
+              result.concat(
+                current.items.map((e) => {
+                  return { ...e, id: e._id };
+                })
+              ),
+            []
+          )}
       />
     </div>
   );
